backend/requirements.txt
fastapi
uvicorn[standard]
pyautogui
websockets
google-generativeai
python-dotenv
